+++ src/util.test.ts
@@ -7,6 +7,7 @@ import test, { ExecutionContext } from "ava";
 import * as sinon from "sinon";
 
 import * as api from "./api-client";
+import { Config, PackWithVersion } from "./config-utils";
 import { getRunnerLogger, Logger } from "./logging";
 import { setupTests } from "./testing-utils";
 import * as util from "./util";
@@ -26,7 +27,7 @@ test("getMemoryFlag() should return the correct --ram flag", (t) => {
   const totalMem = Math.floor(os.totalmem() / (1024 * 1024));
   const expectedThreshold = process.platform === "win32" ? 1536 : 1024;
 
-  const tests = [
+  const tests: Array<[string | undefined, string]> = [
     [undefined, `--ram=${totalMem - expectedThreshold}`],
     ["", `--ram=${totalMem - expectedThreshold}`],
     ["512", "--ram=512"],
@@ -57,7 +58,7 @@ test("getAddSnippetsFlag() should return the correct flag", (t) => {
 test("getThreadsFlag() should return the correct --threads flag", (t) => {
   const numCpus = os.cpus().length;
 
-  const tests = [
+  const tests: Array<[string | undefined, string]> = [
     ["0", "--threads=0"],
     ["1", "--threads=1"],
     [undefined, `--threads=${numCpus}`],
@@ -213,7 +214,10 @@ test("getGitHubVersion", async (t) => {
     auth: "",
     url: "https://ghe.example.com",
   });
-  t.deepEqual({ type: util.GitHubVariant.GHES, version: "2.0" }, v2);
+  t.deepEqual(
+    { type: util.GitHubVariant.GHES, version: "2.0" } as util.GitHubVersion,
+    v2
+  );
 
   mockGetMetaVersionHeader("GitHub AE");
   const ghae = await util.getGitHubVersion({
@@ -287,4 +291,65 @@ async function mockStdInForAuthExpectError(
   await t.throwsAsync(async () =>
     util.getGitHubAuth(mockLogger, undefined, true, stdin)
   );
+}
+
+const ML_POWERED_JS_STATUS_TESTS: Array<[PackWithVersion[], string]> = [
+  [[], "false"],
+  [[{ packName: "someOtherPack" }], "false"],
+  [
+    [{ packName: "someOtherPack" }, util.ML_POWERED_JS_QUERIES_PACK],
+    util.ML_POWERED_JS_QUERIES_PACK.version!,
+  ],
+  [[util.ML_POWERED_JS_QUERIES_PACK], util.ML_POWERED_JS_QUERIES_PACK.version!],
+  [[{ packName: util.ML_POWERED_JS_QUERIES_PACK.packName }], "other"],
+  [
+    [{ packName: util.ML_POWERED_JS_QUERIES_PACK.packName, version: "~0.0.1" }],
+    "other",
+  ],
+  [
+    [
+      { packName: util.ML_POWERED_JS_QUERIES_PACK.packName, version: "0.0.1" },
+      { packName: util.ML_POWERED_JS_QUERIES_PACK.packName, version: "0.0.2" },
+    ],
+    "other",
+  ],
+  [
+    [
+      { packName: "someOtherPack" },
+      { packName: util.ML_POWERED_JS_QUERIES_PACK.packName },
+    ],
+    "other",
+  ],
+];
+
+for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
+  const packDescriptions = `[${packs
+    .map((pack) => JSON.stringify(pack))
+    .join(", ")}]`;
+  test(`ML-powered JS queries status report is "${expectedStatus}" for packs = ${packDescriptions}`, (t) => {
+    return util.withTmpDir(async (tmpDir) => {
+      const config: Config = {
+        languages: [],
+        queries: {},
+        paths: [],
+        pathsIgnore: [],
+        originalUserInput: {},
+        tempDir: tmpDir,
+        toolCacheDir: tmpDir,
+        codeQLCmd: "",
+        gitHubVersion: {
+          type: util.GitHubVariant.DOTCOM,
+        } as util.GitHubVersion,
+        dbLocation: "",
+        packs: {
+          javascript: packs,
+        },
+        debugMode: false,
+        debugArtifactName: util.DEFAULT_DEBUG_ARTIFACT_NAME,
+        debugDatabaseName: util.DEFAULT_DEBUG_DATABASE_NAME,
+      };
+
+      t.is(util.getMlPoweredJsQueriesStatus(config), expectedStatus);
+    });
+  });
 }