+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CodeQL Action and CodeQL Runner Changelog
 
+## 1.0.32 - 07 Feb 2022
+
+- Add `sarif-id` as an output for the `upload-sarif` and `analyze` actions. [#889](https://github.com/github/codeql-action/pull/889)
+- Add `ref` and `sha` inputs to the `analyze` action, which override the defaults provided by the GitHub Action context. [#889](https://github.com/github/codeql-action/pull/889)
+- Update default CodeQL bundle version to 2.8.0. [#911](https://github.com/github/codeql-action/pull/911)
+
 ## 1.0.31 - 31 Jan 2022
 
 - Remove `experimental` message when using custom CodeQL packages. [#888](https://github.com/github/codeql-action/pull/888)